feat(review): introduce aggregation components for candidate grouping, scoring, and result building
